Switch to Synchrony
We've made the switch to using Synchrony, and we've also killed off ModelRocket. Since Synchrony already provides support for Mongoid, I don't see the need to develop our own evented ORM.

Another important change is the switch to Thin and Async Sinatra for HTTP requests. It's still my position that HTTP should only be used to send static files to users, but I have no problem with giving users a more powerful way to accomplish that.