Added gritter and a bit to controllers
I added Gritter, and the redirect example is now shows one of the more interesting possibilities of Rocket.

Notice how the delete action on the Movie controller transmits a message to the user that displays a Gritter notice, then redirects to the index action. Because of the async nature of Rocket, the message is transmitted immediately. Then the redirect occurs, and that information is sent after. It's like pipelining on steroids!